refactor: separate client component for html streaming in server component
