+++ apps/web/src/views/index.ts
@@ -1,6 +1,6 @@
 export { AboutPage } from './about/ui';
 export * from './blog';
-export { GuestbookPage } from './guestbook/ui';
+export { GuestbookContent } from './guestbook/ui';
 export { HomePage } from './home/ui';
 export { LoginPage } from './login/ui';
 export { SpotDetailPage, SpotsPage } from './spots/ui';
